secure_crypto_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pycryptodome==3.20.0
argon2-cffi==23.1.0
python-multipart==0.0.9
aiofiles==23.2.1
